Added a new resource
Added a new resource reference at the bottom - https://www.scaler.com/topics/functional-interface-in-java/. Found this tutorial on online while looking for some resources to learn the functional interface. Topics were explained in detail and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.